gpio/sysfs: Add ID parameter for GPIO lines
On named GPIOs there is currently no find out their numerical id.

Because of this, a GPIO pin named like:
/sys/class/gpio/PROG_B

cannot be unexported by an application different than the one that
exported it.

This patch adds a new parameter to the GPIO line called id, that
shows the numerical id of a given GPIO. E.g.:

$ cat /sys/class/gpio/PROG_B/id
496
$ echo 496 > unexport

Signed-off-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>